Pytorch_Fintuning_Example/Pytorch-Object-Detection-Fintuning-Tutorial-2-main/requirements.txt
absl-py==0.12.0
autopep8==1.5.6
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.23
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
importlib-metadata==4.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.1
numpy==1.20.2
oauthlib==3.1.0
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
toml==0.10.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
